requirements.txt
pandas
numpy
python-dotenv
gradio
gradio-client
pydantic
langchain-community
langchain-chroma
langchain-text-splitters
fastembed
onnxruntime
chromadb
